Implement high-performance Rust core library with WASM compilation support
CORE PERFORMANCE FEATURES:
- Zero-copy JSON parsing with simd-json optimizations
- Memory-efficient compact expression tree representations
- Parallel expression evaluation using rayon thread pools
- SIMD-optimized mathematical operations (add, multiply, dot product)
- Custom bump allocators for large model handling

RUST-SPECIFIC IMPLEMENTATIONS:
- Type-safe CompactExpr with stack-based evaluation
- Error handling with detailed PerformanceError diagnostics
- Iterator-based expression traversal with postfix notation
- Trait-based extensibility for custom operators
- Feature flags for modular compilation

WEBASSEMBLY COMPILATION:
- Enhanced WASM bindings with performance functions
- JavaScript/TypeScript compatible API extensions
- Performance info and benchmarking utilities
- Optimized for browser deployment with feature detection

ADVANCED FEATURES:
- Comprehensive benchmark suite with criterion.rs
- Performance module with parallel stoichiometric matrix computation
- SIMD vector operations with 4-element wide processing
- Custom allocator support with bump allocation
- Zero-copy parsing for improved throughput

Key deliverables implemented:
- Complete performance.rs module with all optimization features
- Enhanced Cargo.toml with performance feature flags
- Comprehensive benchmark infrastructure
- Updated WASM bindings with performance extensions
- Updated README with performance documentation
- Working test suite for all performance features

+++ packages/esm-format-rust/src/reactions.rs
@@ -300,13 +300,13 @@ fn contains_variable(expr: &Expr, var_name: &str) -> bool {
 ///
 /// # Returns
 ///
-/// * `Vec<Vec<i32>>` - Matrix with species as rows and reactions as columns
-pub fn stoichiometric_matrix(system: &ReactionSystem) -> Vec<Vec<i32>> {
+/// * `Vec<Vec<f64>>` - Matrix with species as rows and reactions as columns
+pub fn stoichiometric_matrix(system: &ReactionSystem) -> Vec<Vec<f64>> {
     let num_species = system.species.len();
     let num_reactions = system.reactions.len();
 
     // Initialize matrix with zeros
-    let mut matrix = vec![vec![0i32; num_reactions]; num_species];
+    let mut matrix = vec![vec![0.0f64; num_reactions]; num_species];
 
     // Create mapping from species name to index
     let species_index: HashMap<String, usize> = system.species
@@ -320,7 +320,7 @@ pub fn stoichiometric_matrix(system: &ReactionSystem) -> Vec<Vec<i32>> {
         // Process substrates (negative coefficients)
         for substrate in &reaction.substrates {
             if let Some(&species_idx) = species_index.get(&substrate.species) {
-                let coeff = substrate.coefficient.unwrap_or(1.0) as i32;
+                let coeff = substrate.coefficient.unwrap_or(1.0);
                 matrix[species_idx][reaction_idx] -= coeff;
             }
         }
@@ -328,7 +328,7 @@ pub fn stoichiometric_matrix(system: &ReactionSystem) -> Vec<Vec<i32>> {
         // Process products (positive coefficients)
         for product in &reaction.products {
             if let Some(&species_idx) = species_index.get(&product.species) {
-                let coeff = product.coefficient.unwrap_or(1.0) as i32;
+                let coeff = product.coefficient.unwrap_or(1.0);
                 matrix[species_idx][reaction_idx] += coeff;
             }
         }
@@ -465,13 +465,13 @@ mod tests {
 
         // Check specific values
         // Species A: [-1, 0, -2] (consumed in reactions 1 and 3)
-        assert_eq!(matrix[0], vec![-1, 0, -2]);
+        assert_eq!(matrix[0], vec![-1.0, 0.0, -2.0]);
 
         // Species B: [1, -1, 0] (produced in reaction 1, consumed in reaction 2)
-        assert_eq!(matrix[1], vec![1, -1, 0]);
+        assert_eq!(matrix[1], vec![1.0, -1.0, 0.0]);
 
         // Species C: [0, 1, 1] (produced in reactions 2 and 3)
-        assert_eq!(matrix[2], vec![0, 1, 1]);
+        assert_eq!(matrix[2], vec![0.0, 1.0, 1.0]);
     }
 
     #[test]
